WE NEED TO BE SAFE LIKE HIGH SCHOOL KIDS ON PROM NIGHT
Back in 1983 your mother and I went to a Neil Diamond concert. She drove her
1979 White Pontiac Firebird down I-376 going as fast as her heartbeat. As we
arrived closer and closer to the Civic Arena she kept increasing speed with the
tempo of the slowly quickening music.

Fast and faster, we rode down the street to live out our fantasy of seeing
Leslie in person. As we arrived to the parking lot our passions finally
overflowed, like Iron City Lite on a Hockey Night. We jumped into the tiny back
seat of this middle-class 80s gal sports car as she reached over to change the
8-track to Peace Train by Cat Stevens.

The windows steamed up like the sweat on a Steel Workers brow after a hard days
work. The humidity of this Hot August Night was almost unbearable but this
moment simply could not pass. We had a night to remember, a night I wished would
remain eternal.

"What the fuck was I talking about again?", my dad said.

"Oh right. Use protection. Nine months later you showed up and we had to sell
the Firebird. Fuck you son".